- Removed floating point types from portable example - Minor bug fix

+++ examples/serializable/a_portable_example.cpp
@@ -17,14 +17,14 @@ int main ()
         value <NAME("Field 1"), little_endian <int32_t>>,
 
         // Store the second field using big endian byte ordering
-        value <NAME("Field 2"), big_endian <double>>,
+        value <NAME("Field 2"), big_endian <uint64_t>>,
 
         // Store the wide string as a size-delimited container, storing the size and 
         // characters (see stl_wstring documentation) using little-endian byte ordering.
         value <NAME("Field 3"), stl_wstring <little_endian <uint32_t>>>>;
 
     // Create an object
-    object o1 {1, 2.0, L"Hello World!"};
+    object o1 {1, 2, L"Hello World!"};
 
     // Write the object to a file
     assert(write(o1, std::ofstream("filename")));
